Lecture2/requirements.txt
llama-index-llms-openai
llama-index
python-dotenv
rich
